Refactor the audio fetching testing
