+++ cypress.json
@@ -4,6 +4,6 @@
   "testFiles": "**/*spec.{js,ts}",
   "supportFile": "cypress/support/index.ts",
   "pluginsFile": false,
-  "viewportHeight": 800,
-  "viewportWidth": 720
+  "viewportHeight": 720,
+  "viewportWidth": 360
 }
