--- cypress/fixtures/example.json
@@ -1,3 +1,0 @@
-{
-  "example": "text"
-}
