+++ cypress/fixtures/index.ts
@@ -1,0 +1,1 @@
+export { default as constants } from './constants';
