+++ cypress/support/index.ts
@@ -1,3 +1,1 @@
 import './commands';
-
-export { default as constants } from './constants';
